pyproject.toml
[tool.pytest.ini_options]
minversion = "7.2"
addopts = [
    "-rA"
]
testpaths = [
    "tests",
]

[tool.poetry]
name = "check_done"
version = "1.0.0"
description = "Check done issues on a GitHub project board"
readme = "README.md"
authors = ["Siisurit e.U."]
license = "MIT"
homepage = "https://github.com/siisurit/check_done"
repository = "https://github.com/siisurit/check_done.git"
documentation = "https://github.com/siisurit/check_done/blob/main/README.md"
keywords = ["check", "done", "done done", "issue", "task"]
classifiers = [
        "Development Status :: 4 - Beta",
        "Environment :: Console",
        "Intended Audience :: Developers",
        "License :: OSI Approved :: MIT License",
        "Natural Language :: English",
        "Operating System :: OS Independent",
        "Programming Language :: Python :: 3 :: Only",
        "Programming Language :: Python :: 3.10",
        "Programming Language :: Python :: 3.11",
        "Programming Language :: Python :: 3.12",
        "Programming Language :: Python :: 3.13",
        "Topic :: Software Development",
        "Topic :: Software Development :: Quality Assurance",
]
packages = [
    { include = "check_done" },
]

[tool.poetry.dependencies]
python = ">=3.10, <3.14"  # NOTE We cannot use <4 because of coveralls.
pygments = "^2"
chardet = "^5"
rich = ">=9, <14"
gitpython = "^3"
pyyaml = "^6.0.2"
python-dotenv = "^1.0.1"
pytest-mock = "^3.14.0"
pydantic = "^2.9.2"
pyjwt = "^2.9.0"
requests-mock = "^1.12.1"

[tool.poetry.dev-dependencies]
# coveralls = "^4" TODO: Re-enable once support for python 3.13 is added. See https://github.com/TheKevJames/coveralls-python/pull/542
coverage = "^7"
pytest = "^8"
pytest-cov = "^5"
pre-commit = "^3"
ruff = "^0.5.1"
twine = "^5"

[tool.poetry.scripts]
check_done = "check_done.command:main"

[tool.poetry.urls]
"Issue Tracker" = "https://github.com/siisurit/check_done/issues"

[tool.ruff]
exclude = [
    ".eggs",
    ".git",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".vscode",
    "__pypackages__",
    "_build",
    "build",
    "dist",
    "htmlcov",
]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
ignore = [
    # Missing trailing comma → May cause conflicts when used with the formatter.
    "COM812",
    # Too many branches
    "PLR0912",
    # Too many arguments in function definition
    "PLR0913",
    # Too many statements
    "PLR0915",
    # Magic value used in comparison
    "PLR2004",
    "PTH100",
    "PTH103",
    "PTH107",
    "PTH109",
    "PTH110",
    "PTH112",
    "PTH114",
    "PTH118",
    "PTH119",
    "PTH120",
    "PTH122",
    "PTH123",
    "PTH202",
    "PTH207",
    # Unneccesarry assign → We regularly use `result = ...; return result` to examine the result in the debugger.
    "RET504",
    # Mutable class attributes should be annotated with `typing.ClassVar`
    "RUF012",
    # Avoid specifying long messages outside the exception class
    "TRY003",
    # Abstract `raise` to an inner function
    "TRY301",
]
select = [
    # flake8-builtins
    "A",
    # flake8-bugbear
    "B",
    # flake8-commas
    "COM",
    # flake8-comprehensions
    "C4",
    # flake8-django
    "DJ",
    # flake8-datetimez
    "DTZ",
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # isort
    "I",
    # flake8-no-pep420
    "INP",
    #  flake8-gettext
    "INT",
    # flake8-logging
    "LOG",
    # perflint
    "PERF",
    # pygrep-hooks
    "PGH",
    # flake8-pie
    "PIE",
    # pylint
    "PL",
    # flake8-use-pathlib
    "PTH",
    # refactor
    "R",
    # flake8-raise
    "RSE",
    # flake8-return
    "RET",
    # ruff specific rules
    "RUF",
    # flake8-self
    "SLF",
    # flake8-simplify
    "SIM",
    # tryceratops
    "TRY",
    # flake8-debugger
    "T10",
    # flake8-print
    "T20",
    # pyupgrade
    "UP",
]

[tool.ruff.lint.isort]
known-first-party = ["check_done", "scripts", "tests"]

[build-system]
requires = ["poetry-core>=1.8.1"]
build-backend = "poetry.core.masonry.api"
